requirements.txt
tweepy
python-dotenv
nltk
discord